move video_utils to camera_utils, add base CameraMover class for manipulating camera
